Installer reloads config so create org uses new API endpoint
